Split test_twitter
